requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
authlib==1.6.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bertopic==0.17.3
bleach==6.2.0
brotli==1.1.0
browser-use==0.5.5
bubus==1.4.5
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
cryptography==45.0.5
cython==3.1.2
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.18.0
fqdn==1.5.1
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.26.0
googleapis-common-protos==1.70.0
gradio==5.38.0
gradio-client==1.11.0
greenlet==3.2.3
groovy==0.1.2
groq==0.30.0
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
langchain==0.3.26
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.7
llvmlite==0.44.0
lxml==6.0.0
markdown-it-py==3.0.0
markdown-pdf==1.7
markdownify==1.1.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.12.0
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
narwhals==1.47.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.4
notebook-shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
ollama==0.5.1
openai==1.96.1
orjson==3.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
patchright==1.52.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
playwright==1.53.0
plotly==6.2.0
portalocker==2.10.1
posthog==6.1.1
prometheus-client==0.22.1
prompt-toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pyee==13.0.0
pygments==2.19.2
pymupdf==1.25.3
pynndescent==0.5.13
pyobjc-core==11.1
pyobjc-framework-cocoa==11.1
pyparsing==3.2.3
pypdf==5.8.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
screeninfo==0.8.1
semantic-version==2.10.0
send2trash==1.8.3
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.1
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
umap-learn==0.5.9.post2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
zstandard==0.23.0
